Fixed int to integer type in tables

+++ site/database/migrations/2023_04_04_091055_create_room_prices_table.php
@@ -20,8 +20,8 @@ class CreateRoomPricesTable extends Migration
             //room prices information
             $table->boolean('view');
             $table->string('type');
-            $table->float('price');
-            $table->int('capacity');
+            $table->integer('price');
+            $table->integer('capacity');
         });
     }
